+++ motifstudio-web/src/app/useDebounce.tsx
@@ -1,0 +1,19 @@
+"use client";
+import { useState } from "react";
+import { useEffect } from "react";
+
+export function useDebounce(value: any, delay: number) {
+    const [debouncedValue, setDebouncedValue] = useState(value);
+
+    useEffect(() => {
+        const handler = setTimeout(() => {
+            setDebouncedValue(value);
+        }, delay);
+
+        return () => {
+            clearTimeout(handler);
+        };
+    }, [value, delay]);
+
+    return debouncedValue;
+}
